refactor: extract poll and lottery sections
